convert `WorkflowRename` to a generic async `RenameModal` component
