+++ client/src/components/Common/RenameModal.test.ts
@@ -1,4 +1,4 @@
-import { getLocalVue, suppressBootstrapVueWarnings } from "@tests/vitest/helpers";
+import { getLocalVue } from "@tests/vitest/helpers";
 import { createWrapper, mount } from "@vue/test-utils";
 import flushPromises from "flush-promises";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
@@ -7,7 +7,7 @@ import type { WorkflowSummary } from "@/api/workflows";
 import { updateWorkflow } from "@/components/Workflow/workflows.services";
 import { Toast } from "@/composables/toast";
 
-import WorkflowRename from "./WorkflowRename.vue";
+import RenameModal from "./RenameModal.vue";
 import GModal from "@/components/BaseComponents/GModal.vue";
 
 vi.mock("@/components/Workflow/workflows.services", () => ({
@@ -24,19 +24,22 @@ const localVue = getLocalVue();
 const WORKFLOW_ID = "workflow-abc123";
 const WORKFLOW_NAME = "My Test Workflow";
 
-async function mountWorkflowRename(name = WORKFLOW_NAME) {
-    const wrapper = mount(WorkflowRename as object, {
+async function mountRenameModal(name = WORKFLOW_NAME) {
+    const wrapper = mount(RenameModal as object, {
         localVue,
-        propsData: { id: WORKFLOW_ID, name },
+        propsData: {
+            name,
+            itemType: "workflow",
+            renameAction: (newName: string) => updateWorkflow(WORKFLOW_ID, { name: newName }),
+        },
         attachTo: document.body,
     });
     await flushPromises();
     return wrapper;
 }
 
-describe("WorkflowRename", () => {
+describe("RenameModal tested for renaming workflows", () => {
     beforeEach(() => {
-        suppressBootstrapVueWarnings();
         vi.clearAllMocks();
     });
 
@@ -47,7 +50,7 @@ describe("WorkflowRename", () => {
     it("calls updateWorkflow with the new name and emits close on confirm", async () => {
         vi.mocked(updateWorkflow).mockResolvedValue({} as WorkflowSummary);
 
-        const wrapper = await mountWorkflowRename();
+        const wrapper = await mountRenameModal();
 
         await createWrapper(document.body).find("#workflow-name-input").setValue("Renamed Workflow");
         wrapper.findComponent(GModal).vm.$emit("ok");
@@ -60,7 +63,7 @@ describe("WorkflowRename", () => {
     it("shows toast error and emits close when update fails, preserving original name on reopen", async () => {
         vi.mocked(updateWorkflow).mockRejectedValue(new Error("Server error"));
 
-        const wrapper = await mountWorkflowRename();
+        const wrapper = await mountRenameModal();
 
         await createWrapper(document.body).find("#workflow-name-input").setValue("Attempted New Name");
         wrapper.findComponent(GModal).vm.$emit("ok");
@@ -73,14 +76,14 @@ describe("WorkflowRename", () => {
 
         // Simulate parent closing and reopening the modal (destroys old instance)
         wrapper.destroy();
-        await mountWorkflowRename();
+        await mountRenameModal();
         expect((createWrapper(document.body).find("#workflow-name-input").element as HTMLInputElement).value).toBe(
             WORKFLOW_NAME,
         );
     });
 
     it("keeps ok button disabled when name is unchanged", async () => {
-        const wrapper = await mountWorkflowRename();
+        const wrapper = await mountRenameModal();
 
         expect(wrapper.findComponent(GModal).props("okDisabled")).toBe(true);
     });
